+++ project-loader/src/main/kotlin/de/itemis/mps/gradle/logging/LogLevel.kt
@@ -1,9 +1,0 @@
-package de.itemis.mps.gradle.logging
-
-public enum class LogLevel {
-    ALL,
-    INFO,
-    WARN,
-    ERROR,
-    OFF
-}